# src/Swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "CAPTION CUỐI KHÓA"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getAllUsers",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getUserById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "updateUserById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateUserByIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "deleteUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/users/get-user-by-page": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getUserByPage",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageIndex",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/users/get-user-by-token": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getUserByToken",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/users/search/{name}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getUserByName",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "definitions": {
...
    "updateUserByIdRequest": {
      "title": "updateUserByIdRequest",
      "example": {
        "ho_ten": "hoàng",
        "email": "hhoaang@gmail.com",
        "mat_khau": "1234",
        "dien_thoai": "012345679",
        "ngay_sinh": "1990-01-22",
        "gioi_tinh_id": 1,
        "role_id": 1
      },
      "type": "object",
      "properties": {
        "ho_ten": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "mat_khau": {
          "type": "string"
        },
        "dien_thoai": {
          "type": "string"
        },
        "ngay_sinh": {
          "type": "string"
        },
        "gioi_tinh_id": {
          "type": "integer",
          "format": "int32"
        },
        "role_id": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "ho_ten",
        "email",
        "mat_khau",
        "dien_thoai",
        "ngay_sinh",
        "gioi_tinh_id",
        "role_id"
      ]
    },
...
  },
  "security": [],
...
}